TomatoLeafs/requirements.txt
Flask==2.2.2
Flask_WTF==1.0.1
numpy==1.18.5
opencv_python==4.4.0.44
Pillow==9.4.0
pytest==7.2.1
WTForms==3.0.1
